+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 <!-- There is always Unreleased section on the top. Subsections (Added, Changed, Fixed, Removed) should be added as needed. -->
 
 ## Unreleased
+### Added
+- Commands which include user now implements `UserAwareInterface` and `getUserId()` method (ie. Interaction, Sorting, UserMerge, UserRecommendation).
+
+### Changed
+- Validate all commands of `recommendation()` and `sorting()` request involve the same user.
 
 ## 0.10.0 - 2017-11-30
 ### Changed